Improves efficiency of spawning tasks in the executor, by pre-computing the task queue index

Before:
```
cost to run task no task queue: 126ns
cost to run task in task queue: 300ns
```

After
```
cost to run task no task queue: 93ns
cost to run task in task queue: 257ns
```

No regressions found in any of the existing benchmarks. Semaphore and local channel show similar improvements for the contended case.