@glommer 2020-12-03
Thanks @zserik . I will leave this patch cooking for a bit longer. Due to a change that I consider too drastic (the fact that you can no longer spawn before run), I will bump the version to 0.3.0 once I do it.