@zserik 2020-12-04
I created a PR to `glommer:executor` which contains a better suggestion, as I've made an error in the last one...