@glommer 2020-12-08
ETA to publish a new crate is ~ 2 weeks, hopefully when the network API stabilizes